Add right-side offset for + button in TagButtons
Adds a new parameter to the Draw call in TagButtons.cs that allows for an offset
to prevent the + button from being drawn within a certain distance of the right
edge of the screen.

This is done to allow elements in different frames to be aligned with the +
button vertically without overlapping with it.